renamed Error to exception
- renamed error to exception for consistency
- refactored paths referencing error to exception
