+++ src/main/kotlin/dev/seekheart/jobtrackerapi/handlers/GlobalErrorController.kt
@@ -1,7 +1,7 @@
 package dev.seekheart.jobtrackerapi.handlers
 
-import dev.seekheart.jobtrackerapi.users.errors.UserAlreadyExistsError
-import dev.seekheart.jobtrackerapi.users.errors.UserNotFoundException
+import dev.seekheart.jobtrackerapi.users.exceptions.UserAlreadyExistsException
+import dev.seekheart.jobtrackerapi.users.exceptions.UserNotFoundException
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.ResponseStatus
 @ControllerAdvice
 class GlobalErrorController {
     @ResponseStatus(HttpStatus.CONFLICT)
-    @ExceptionHandler(UserAlreadyExistsError::class)
+    @ExceptionHandler(UserAlreadyExistsException::class)
     fun handleAlreadyExistsErrors() {
     }
