--- src/main/kotlin/dev/seekheart/jobtrackerapi/users/errors/UserAlreadyExistsError.kt
@@ -1,4 +1,0 @@
-package dev.seekheart.jobtrackerapi.users.errors
-
-class UserAlreadyExistsError(name: String) : IllegalArgumentException("User name = $name already exists!") {
-}
