+++ src/main/kotlin/dev/seekheart/jobtrackerapi/users/exceptions/UserAlreadyExistsException.kt
@@ -1,0 +1,4 @@
+package dev.seekheart.jobtrackerapi.users.exceptions
+
+class UserAlreadyExistsException(name: String) : IllegalArgumentException("User name = $name already exists!") {
+}
